Improve org & dataset selection robustness
There was a little bug when we were selecting a dataset, then turn
custom dataset selection off, then change organism, and then go back to
the previously selected organism and turn custom data selection on. The
dataset was not loading because some variable were set to null and not
re-computed.